requirements.txt
python-whois
cryptography
dnspython
aiohttp
idna
